Small fixes/improvements found by using clang. Unfortunatly, we cant't really use clang yet because init-priorites don't work as they should.

+++ nre/include/util/Atomic.h
@@ -75,8 +75,8 @@ public:
 		// We don't need a lock prefix, because this is only meant to be
 		// uninterruptible not atomic. So don't use this for SMP!
 		asm ("cmpxchg8b %2\n"
-				: "+a" (static_cast<uint32_t>(lo)), "+d" (static_cast<uint32_t>(hi))
-				: "m" (v), "b" (static_cast<uint32_t>(lo)), "c" (static_cast<uint32_t>(hi)));
+				: "+a" (lo), "+d" (hi)
+				: "m" (v), "b" (lo), "c" (hi));
 		return (static_cast<uint64_t>(hi) << 32) | lo;
 	}
 
@@ -92,8 +92,8 @@ public:
 		asm ("1: cmpxchg8b %0\n"
 				"jne 1b"
 				: "+m" (to),
-				"+a" (static_cast<uint32_t>(olo)), "+d" (static_cast<uint32_t>(ohi))
-				: "b" (static_cast<uint32_t>(nlo)), "c" (static_cast<uint32_t>(nhi)));
+				"+a" (olo), "+d" (ohi)
+				: "b" (nlo), "c" (nhi));
 	}
 
 private:
